:sparkles: Connecting REST API with Proposal Flow
Refactoring packages duo problems with deployNodes, spliting in 3 projects
