+++ .gitignore
@@ -2,6 +2,7 @@ HELP.md
 .gradle
 .gradletasknamecache
 build/
+logs/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**
 !**/src/test/**
