+++ application/cordapp-contracts-states/build.gradle
@@ -1,0 +1,10 @@
+apply plugin: "net.corda.plugins.cordapp"
+apply plugin: "net.corda.plugins.cordformation"
+apply plugin: "net.corda.plugins.quasar-utils"
+
+dependencies {
+
+    cordaCompile libs.corda
+
+    cordaRuntime "$cordaGroup:corda:$cordaVersion"
+}
