+++ application/cordapp-flows/build.gradle
@@ -1,0 +1,12 @@
+apply plugin: "net.corda.plugins.cordapp"
+apply plugin: "net.corda.plugins.cordformation"
+apply plugin: "net.corda.plugins.quasar-utils"
+
+dependencies {
+
+    cordapp project(":application:cordapp-contracts-states")
+
+    cordaCompile libs.corda
+
+    cordaRuntime "$cordaGroup:corda:$cordaVersion"
+}
