+++ application/rest-api/plugins/mapstruct.gradle
