+++ application/rest-api/plugins/openapi.gradle
@@ -1,0 +1,42 @@
+buildscript {
+
+    repositories {
+        mavenLocal()
+        maven { url "https://repo1.maven.org/maven2" }
+    }
+
+    dependencies {
+        classpath "org.openapitools:openapi-generator-gradle-plugin:$openAPIGeneratorVersion"
+    }
+}
+
+apply plugin: "org.openapi.generator"
+
+def cambioApiServerOutput = "$buildDir/generated/openapi-code-server".toString()
+
+task generateCambioApiServer(type: org.openapitools.generator.gradle.plugin.tasks.GenerateTask) {
+
+    generatorName = "spring"
+    inputSpec = "$projectDir/src/main/resources/openapi/cambio-api.yaml".toString()
+    outputDir = cambioApiServerOutput
+
+    apiPackage = "org.openapi.cambio.server.api"
+    modelPackage = "org.openapi.cambio.server.model"
+
+    configOptions = [
+        "dateLibrary": "java8",
+        "hideGenerationTimestamp": "true",
+        "interfaceOnly": "true",
+        "java8": "true",
+        "performBeanValidation": "false",
+        "serializableModel": "true",
+        "useBeanValidation": "false",
+        "useOptional": "true"
+    ]
+}
+
+compileKotlin.dependsOn(
+    generateCambioApiServer
+)
+
+sourceSets.main.java.srcDir "$cambioApiServerOutput/src/main/java"
