+++ application/rest-api/src/main/kotlin/br/com/fiap/mba/mscambio/exceptions/PropostaInvalidaException.kt
