+++ application/rest-api/src/main/kotlin/br/com/fiap/mba/mscambio/gateways/NodeRPCConnection.kt
@@ -1,0 +1,33 @@
+package br.com.fiap.mba.mscambio.gateways
+
+import net.corda.client.rpc.CordaRPCClient
+import net.corda.client.rpc.CordaRPCConnection
+import net.corda.core.messaging.CordaRPCOps
+import net.corda.core.utilities.NetworkHostAndPort
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.stereotype.Component
+import javax.annotation.PreDestroy
+
+@Component
+open class NodeRPCConnection(
+    @Value("\${config.rpc.host}") private val host: String,
+    @Value("\${config.rpc.port}") private val rpcPort: Int,
+    @Value("\${config.rpc.username}") private val username: String,
+    @Value("\${config.rpc.password}") private val password: String
+) : AutoCloseable {
+
+    private val rpcConnection: CordaRPCConnection
+    final val proxy: CordaRPCOps
+
+    init {
+        val rpcAddress = NetworkHostAndPort(host, rpcPort)
+        val rpcClient = CordaRPCClient(rpcAddress)
+        rpcConnection = rpcClient.start(username, password)
+        proxy = rpcConnection.proxy
+    }
+
+    @PreDestroy
+    override fun close() {
+        rpcConnection.notifyServerAndClose()
+    }
+}
