+++ application/rest-api/src/main/resources/application-dev.yaml
