+++ application/rest-api/src/main/resources/application.yaml
