+++ application/rest-api/src/main/resources/openapi/cambio-api.yaml
